拆分widget.js文件，只留父类
