+++ src/views/partComponent/components/scrollBoard/ScrollBoardChildOne/index.vue
