+++ src/views/partComponent/components/scrollBoard/widget_class.js
@@ -1,0 +1,82 @@
+import { ScrollBoard } from '../../widget'
+import widgetFields from '../../widget_fields' // 单一的form
+// import validators from '../../validators'
+import validatorCombine from '../../validator_combine'
+import widgetFieldsCombine from '../../widget_fields_combine'
+import jsonTemplate from '../../json_template'
+
+class ScrollBoardChild extends ScrollBoard {
+  static componentDescription = {
+    name: '轮播表一',
+    parent: 'ScrollBoard',
+    img: 'scrollboard.png'
+  }
+  constructor(constructorData) {
+    super(constructorData)
+    this.componentKey = 'ScrollBoardChild'
+
+    this.setStyleFields(constructorData, {
+      headerBGC: widgetFields.FieldColorPicker({
+        label: '表头背景色',
+        formModel: '#00BAFF'
+      }),
+      oddRowBGC: widgetFields.FieldColorPicker({
+        label: '奇数行背景色',
+        formModel: '#003B51'
+      }),
+      evenRowBGC: widgetFields.FieldColorPicker({
+        label: '偶数行背景色',
+        formModel: '#0A2732'
+      }),
+      headerColor: widgetFields.FieldColorPicker({
+        label: '表头文字颜色',
+        formModel: '#FFF'
+      }),
+      bodyColor: widgetFields.FieldColorPicker({
+        label: '主体文字颜色',
+        formModel: '#FFF'
+      }),
+      headerFontSize: widgetFields.FiledSlider({
+        label: '表头字号',
+        formModel: 16,
+        max: 50,
+        min: 10
+      }),
+      bodyFontSize: widgetFields.FiledSlider({
+        label: '主体字号',
+        formModel: 14,
+        max: 50,
+        min: 10
+      }),
+      index: widgetFields.FiledSwitch({
+        label: '显示行号',
+        formModel: false
+      }),
+      rowNum: widgetFields.FiledSlider({
+        label: '显示表行数',
+        formModel: 5,
+        max: 20,
+        min: 1
+      })
+    })
+
+    this.setFields(constructorData, {
+      isFakeData: widgetFieldsCombine.FieldsDataSource({
+        validator: [validatorCombine.validateFieldDataSource()],
+        fakeData: JSON.stringify(jsonTemplate.ForScrollBoard),
+        radios: [
+          {
+            label: '使用模拟数据',
+            value: 'fake'
+          },
+          {
+            label: '使用接口数据',
+            value: 'real'
+          }
+        ]
+      })
+    })
+  }
+}
+
+export { ScrollBoardChild }
