+++ src/views/partComponent/widget.js
@@ -1,9 +1,9 @@
 import _ from 'lodash'
-import widgetFields from './widget_fields' // 单一的form
-import widgetFieldsCombine from './widget_fields_combine' // 定制化的form
-import validators from './validators'
-import validatorCombine from './validator_combine'
-import jsonTemplate from './json_template'
+// import widgetFields from './widget_fields' // 单一的form
+// import widgetFieldsCombine from './widget_fields_combine' // 定制化的form
+// import validators from './validators'
+// import validatorCombine from './validator_combine'
+// import jsonTemplate from './json_template'
 
 /**
  * uuid: 前端唯一标识
@@ -93,85 +93,6 @@ class Title extends Widget {
   }
 }
 
-class TitleChildOne extends Title {
-  static componentDescription = {
-    name: '子标题1',
-    parent: 'Title',
-    img: 'TitleChildOne_icon.jpg'
-  }
-  constructor(constructorData) {
-    super(constructorData)
-    this.componentKey = 'TitleChildOne'
-
-    this.setFields(constructorData, {
-      url: widgetFields.FieldInput({
-        label: '链接',
-        validator: [validators.requireInput('请输入链接')]
-      })
-    })
-
-    this.setStyleFields(constructorData, {
-      color: widgetFields.FieldColorPicker({
-        label: '颜色',
-        formModel: '#000'
-      }),
-      fontSize: widgetFields.FieldSelect({
-        label: '字体大小',
-        options: ['12px', '20px', '30px']
-      })
-    })
-  }
-}
-
-class TitleChildTwo extends Title {
-  static componentDescription = {
-    name: '子标题2',
-    parent: 'Title',
-    img: 'TitleChildOne_icon.jpg'
-  }
-  constructor(constructorData) {
-    super(constructorData)
-
-    this.componentKey = 'TitleChildTwo'
-
-    this.setFields(constructorData, {
-      url: widgetFields.FieldInput({
-        label: '链接',
-        validator: [validators.requireInput('请输入链接')]
-      }),
-      ip: widgetFields.FieldInput({
-        label: '地址',
-        validator: [validators.requireInput('请输入地址')]
-      }),
-      isFakeData: widgetFieldsCombine.FieldsDataSource({
-        validator: [validatorCombine.validateFieldDataSource()],
-        fakeData: JSON.stringify(jsonTemplate.ForTestComponentThree),
-        radios: [
-          {
-            label: '使用接口数据',
-            value: 'real'
-          },
-          {
-            label: '使用模拟数据',
-            value: 'fake'
-          }
-        ]
-      })
-    })
-
-    this.setStyleFields(constructorData, {
-      color: widgetFields.FieldColorPicker({
-        label: '颜色',
-        formModel: '#000'
-      }),
-      fontSize: widgetFields.FieldSelect({
-        label: '字体大小',
-        options: ['12px', '20px', '30px']
-      })
-    })
-  }
-}
-
 class ScrollBoard extends Widget {
   static componentDescription = {
     name: '轮播表类',
@@ -187,201 +108,4 @@ class ScrollBoard extends Widget {
   }
 }
 
-class ScrollBoardChild extends ScrollBoard {
-  static componentDescription = {
-    name: '轮播表一',
-    parent: 'ScrollBoard',
-    img: 'scrollboard.png'
-  }
-  constructor(constructorData) {
-    super(constructorData)
-    this.componentKey = 'ScrollBoardChild'
-
-    this.setStyleFields(constructorData, {
-      headerBGC: widgetFields.FieldColorPicker({
-        label: '表头背景色',
-        formModel: '#00BAFF'
-      }),
-      oddRowBGC: widgetFields.FieldColorPicker({
-        label: '奇数行背景色',
-        formModel: '#003B51'
-      }),
-      evenRowBGC: widgetFields.FieldColorPicker({
-        label: '偶数行背景色',
-        formModel: '#0A2732'
-      }),
-      headerColor: widgetFields.FieldColorPicker({
-        label: '表头文字颜色',
-        formModel: '#FFF'
-      }),
-      bodyColor: widgetFields.FieldColorPicker({
-        label: '主体文字颜色',
-        formModel: '#FFF'
-      }),
-      headerFontSize: widgetFields.FiledSlider({
-        label: '表头字号',
-        formModel: 16,
-        max: 50,
-        min: 10
-      }),
-      bodyFontSize: widgetFields.FiledSlider({
-        label: '主体字号',
-        formModel: 14,
-        max: 50,
-        min: 10
-      }),
-      index: widgetFields.FiledSwitch({
-        label: '显示行号',
-        formModel: false
-      }),
-      rowNum: widgetFields.FiledSlider({
-        label: '显示表行数',
-        formModel: 5,
-        max: 20,
-        min: 1
-      })
-    })
-
-    this.setFields(constructorData, {
-      isFakeData: widgetFieldsCombine.FieldsDataSource({
-        validator: [validatorCombine.validateFieldDataSource()],
-        fakeData: JSON.stringify(jsonTemplate.ForScrollBoard),
-        radios: [
-          {
-            label: '使用模拟数据',
-            value: 'fake'
-          },
-          {
-            label: '使用接口数据',
-            value: 'real'
-          }
-        ]
-      })
-    })
-  }
-}
-
-// class TestComponentOne extends Widget {
-//   static initSize = {
-//     width: 300,
-//     height: 300
-//   }
-
-//   static componentDescription = {
-//     name: '测试组件1',
-//     parent
-//   }
-
-//   constructor(constructorData) {
-//     super(constructorData)
-
-//     this.componentKey = 'TestComponentOne'
-
-//     this.setDragSize(constructorData, {
-//       width: TestComponentOne.initSize.width,
-//       height: TestComponentOne.initSize.height
-//     })
-
-//     this.setFields(constructorData, {
-//       url: widgetFields.FieldInput({
-//         label: '链接',
-//         validator: [validators.requireInput('请输入链接')]
-//       }),
-//       ip: widgetFields.FieldInput({
-//         label: '地址',
-//         validator: [validators.requireInput('请输入地址')]
-//       }),
-//       people: widgetFields.FieldRadio({
-//         validator: [validators.requireInput('请选择人')],
-//         radios: [
-//           {
-//             label: '张三',
-//             value: 'zhangsan'
-//           },
-//           {
-//             label: '李四',
-//             value: 'lisi'
-//           }
-//         ]
-//       })
-//     })
-//   }
-// }
-
-// class TestComponentTwo extends Widget {
-//   static initSize = {
-//     width: 200,
-//     height: 500
-//   }
-//   constructor(constructorData) {
-//     super(constructorData)
-
-//     this.componentKey = 'TestComponentTwo'
-
-//     this.setDragSize(constructorData, {
-//       width: TestComponentTwo.initSize.width,
-//       height: TestComponentTwo.initSize.height
-//     })
-
-//     this.setStyleFields(constructorData, {
-//       color: widgetFields.FieldColorPicker(),
-//       fontSize: widgetFields.FieldSelect({
-//         label: '请选择字体大小',
-//         options: ['12px', '20px', '30px']
-//       })
-//     })
-//   }
-// }
-
-// class TestComponentThree extends Widget {
-//   static initSize = {
-//     width: 200,
-//     height: 500
-//   }
-//   constructor(constructorData) {
-//     super(constructorData)
-
-//     this.componentKey = 'TestComponentThree'
-
-//     this.setDragSize(constructorData, {
-//       width: TestComponentThree.initSize.width,
-//       height: TestComponentThree.initSize.height
-//     })
-
-//     this.setFields(constructorData, {
-//       url: widgetFields.FieldInput({
-//         label: '链接',
-//         validator: [validators.requireInput('请输入链接')]
-//       }),
-//       ip: widgetFields.FieldInput({
-//         label: '地址',
-//         validator: [validators.requireInput('请输入地址')]
-//       }),
-//       isFakeData: widgetFieldsCombine.FieldsDataSource({
-//         validator: [validatorCombine.validateFieldDataSource()],
-//         fakeData: JSON.stringify(jsonTemplate.ForTestComponentThree),
-//         radios: [
-//           {
-//             label: '使用接口数据',
-//             value: 'real'
-//           },
-//           {
-//             label: '使用模拟数据',
-//             value: 'fake'
-//           }
-//         ]
-//       })
-//     })
-
-//     this.setStyleFields(constructorData, {
-//       color: widgetFields.FieldColorPicker(),
-//       fontSize: widgetFields.FieldSelect({
-//         label: '请选择字体大小',
-//         options: ['12px', '20px', '30px']
-//       })
-//     })
-//   }
-// }
-// TestComponentOne, TestComponentTwo, TestComponentThree,
-
-export { Title, TitleChildOne, TitleChildTwo, ScrollBoard, ScrollBoardChild }
+export { Title, ScrollBoard }
